fix auth fix extension path
